+++ backend-rust/src/graphql_api/baker.rs
@@ -1,11 +1,16 @@
 use super::{
-    account::Account, get_config, get_pool, transaction::Transaction, ApiError, ApiResult,
-    ApiServiceConfig, ConnectionQuery,
+    account::Account,
+    baker_and_delegator_types::{DelegationSummary, PaydayPoolReward},
+    get_config, get_pool,
+    transaction::Transaction,
+    ApiError, ApiResult, ApiServiceConfig, ConnectionQuery,
 };
 use crate::{
-    address::AccountAddress,
     connection::{ConcatCursor, ConnectionBounds, DescendingI64, Reversed},
-    graphql_api::node_status::{NodeInfoReceiver, NodeStatus},
+    graphql_api::{
+        node_status::{NodeInfoReceiver, NodeStatus},
+        todo_api,
+    },
     scalar_types::{Amount, BakerId, DateTime, Decimal, MetadataUrl},
     transaction_event::{baker::BakerPoolOpenStatus, Event},
     transaction_reject::TransactionRejectReason,
@@ -113,8 +118,8 @@ impl QueryBaker {
                 )
                 .await
             }
-            BakerSort::BakerApy30DaysDesc => todo!(),
-            BakerSort::DelegatorApy30DaysDesc => todo!(),
+            BakerSort::BakerApy30DaysDesc => todo_api!(),
+            BakerSort::DelegatorApy30DaysDesc => todo_api!(),
             BakerSort::BlockCommissionsAsc => {
                 Baker::block_commission_asc_connection(
                     config,
@@ -2246,10 +2251,6 @@ struct BakerPool<'a> {
     open_status: Option<BakerPoolOpenStatus>,
     metadata_url: Option<&'a str>,
     // TODO: apy(period: ApyPeriod!): PoolApy!
-    // TODO: poolRewards("Returns the first _n_ elements from the list." first: Int "Returns the
-    // elements in the list that come after the specified cursor." after: String "Returns the last
-    // _n_ elements from the list." last: Int "Returns the elements in the list that come before
-    // the specified cursor." before: String): PaydayPoolRewardConnection
 }
 
 #[Object]
@@ -2281,6 +2282,86 @@ impl<'a> BakerPool<'a> {
     async fn open_status(&self) -> Option<BakerPoolOpenStatus> { self.open_status }
 
     async fn metadata_url(&self) -> Option<&'a str> { self.metadata_url }
+
+    async fn pool_rewards(
+        &self,
+        ctx: &Context<'_>,
+        #[graphql(desc = "Returns the first _n_ elements from the list.")] first: Option<u64>,
+        #[graphql(desc = "Returns the elements in the list that come after the specified cursor.")]
+        after: Option<String>,
+        #[graphql(desc = "Returns the last _n_ elements from the list.")] last: Option<u64>,
+        #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
+        before: Option<String>,
+    ) -> ApiResult<connection::Connection<DescendingI64, PaydayPoolReward>> {
+        let pool = get_pool(ctx)?;
+        let config = get_config(ctx)?;
+        let query = ConnectionQuery::<DescendingI64>::new(
+            first,
+            after,
+            last,
+            before,
+            config.pool_rewards_connection_limit,
+        )?;
+        let mut row_stream = sqlx::query_as!(
+            PaydayPoolReward,
+            "SELECT * FROM (
+                SELECT
+                    payday_block_height as block_height,
+                    slot_time,
+                    pool_owner,
+                    payday_total_transaction_rewards as total_transaction_rewards,
+                    payday_delegators_transaction_rewards as delegators_transaction_rewards,
+                    payday_total_baking_rewards as total_baking_rewards,
+                    payday_delegators_baking_rewards as delegators_baking_rewards,
+                    payday_total_finalization_rewards as total_finalization_rewards,
+                    payday_delegators_finalization_rewards as delegators_finalization_rewards
+                FROM bakers_payday_pool_rewards
+                    JOIN blocks ON blocks.height = payday_block_height
+                WHERE pool_owner_for_primary_key = $5 
+                    AND payday_block_height > $2 AND payday_block_height < $1
+                ORDER BY
+                    (CASE WHEN $4 THEN payday_block_height END) ASC,
+                    (CASE WHEN NOT $4 THEN payday_block_height END) DESC
+                LIMIT $3
+                ) AS rewards
+            ORDER BY rewards.block_height DESC",
+            i64::from(query.from),
+            i64::from(query.to),
+            query.limit,
+            query.is_last,
+            self.id
+        )
+        .fetch(pool);
+
+        let mut connection = connection::Connection::new(false, false);
+        while let Some(rewards) = row_stream.try_next().await? {
+            connection.edges.push(connection::Edge::new(rewards.block_height.into(), rewards));
+        }
+
+        if let (Some(edge_min_index), Some(edge_max_index)) =
+            (connection.edges.last(), connection.edges.first())
+        {
+            let result = sqlx::query!(
+                "
+                    SELECT 
+                        MIN(payday_block_height) as min_index,
+                        MAX(payday_block_height) as max_index
+                    FROM bakers_payday_pool_rewards
+                    WHERE pool_owner_for_primary_key = $1
+                ",
+                self.id
+            )
+            .fetch_one(pool)
+            .await?;
+
+            connection.has_previous_page =
+                result.max_index.map_or(false, |db_max| db_max > edge_max_index.node.block_height);
+            connection.has_next_page =
+                result.min_index.map_or(false, |db_min| db_min < edge_min_index.node.block_height);
+        }
+
+        Ok(connection)
+    }
 
     async fn delegators(
         &self,
@@ -2347,31 +2428,6 @@ impl<'a> BakerPool<'a> {
     }
 }
 
-struct DelegationSummary {
-    index:            i64,
-    account_address:  AccountAddress,
-    staked_amount:    i64,
-    restake_earnings: Option<bool>,
-}
-
-#[Object]
-impl DelegationSummary {
-    async fn account_address(&self) -> &AccountAddress { &self.account_address }
-
-    async fn staked_amount(&self) -> ApiResult<Amount> {
-        self.staked_amount.try_into().map_err(|_| {
-            ApiError::InternalError(
-                "Staked amount in database should be a valid UnsignedLong".to_string(),
-            )
-        })
-    }
-
-    async fn restake_earnings(&self) -> ApiResult<bool> {
-        self.restake_earnings.ok_or(ApiError::InternalError(
-            "Delegator should have a boolean in the `restake_earnings` variable.".to_string(),
-        ))
-    }
-}
 #[derive(SimpleObject)]
 struct CommissionRates {
     transaction_commission:  Option<Decimal>,