+++ backend-rust/src/indexer.rs
@@ -133,8 +133,11 @@ impl IndexerService {
     ) -> anyhow::Result<Self> {
         let last_height_stored = sqlx::query!(
             "
-SELECT height FROM blocks ORDER BY height DESC LIMIT 1
-"
+                SELECT height 
+                FROM blocks 
+                ORDER BY height 
+                DESC LIMIT 1
+            "
         )
         .fetch_optional(&pool)
         .await?
@@ -532,14 +535,14 @@ impl BlockProcessor {
     ) -> anyhow::Result<Self> {
         let last_finalized_block = sqlx::query!(
             "
-SELECT
-  hash,
-  cumulative_finalization_time
-FROM blocks
-WHERE finalization_time IS NOT NULL
-ORDER BY height DESC
-LIMIT 1
-"
+            SELECT
+                hash,
+                cumulative_finalization_time
+            FROM blocks
+            WHERE finalization_time IS NOT NULL
+            ORDER BY height DESC
+            LIMIT 1
+            "
         )
         .fetch_one(&pool)
         .await
@@ -547,13 +550,13 @@ LIMIT 1
 
         let last_block = sqlx::query!(
             "
-SELECT
-  slot_time,
-  cumulative_num_txs
-FROM blocks
-ORDER BY height DESC
-LIMIT 1
-"
+            SELECT
+                slot_time,
+                cumulative_num_txs
+            FROM blocks
+            ORDER BY height DESC
+            LIMIT 1
+            "
         )
         .fetch_one(&pool)
         .await
@@ -4258,19 +4261,242 @@ impl PreparedSpecialTransactionOutcomes {
     }
 }
 
+/// The `SpecialEvents` of a payday block in the order they
+/// occur in the block.
+struct PreparedPaydaySpecialTransacionOutcomes {
+    /// Height of the payday block containing the events.
+    block_height: i64,
+    // Total rewards
+    total_rewards_pool_owners: Vec<Option<i64>>,
+    total_transaction_rewards: Vec<i64>,
+    total_baking_rewards: Vec<i64>,
+    total_finalization_rewards: Vec<i64>,
+    // Delegator rewards
+    delegators_rewards_pool_owners: Vec<Option<i64>>,
+    delegators_rewards_canonical_addresses: Vec<Vec<u8>>,
+    delegators_transaction_rewards: Vec<i64>,
+    delegators_baking_rewards: Vec<i64>,
+    delegators_finalization_rewards: Vec<i64>,
+}
+
+impl PreparedPaydaySpecialTransacionOutcomes {
+    fn prepare(block_height: i64, events: &[SpecialTransactionOutcome]) -> anyhow::Result<Self> {
+        // Extract the rewards from the `SpecialEvents` in each payday block
+        // and associate it with the `pool_owner`.
+        // The `pool_owner` can be either a `baker_id` or `None`.
+        // The `pool_owner` is `None` if the pool rewards are for the passive
+        // delegators which can happen at most once per payday block.
+        //
+        // https://docs.rs/concordium-rust-sdk/6.0.0/concordium_rust_sdk/types/enum.SpecialTransactionOutcome.html#variant.PaydayAccountReward
+        // The order of `SpecialEvents` in each payday block has a meaning to
+        // determine which rewards go to the baker of a baker pool and which
+        // rewards go to the pool's delegators.
+        //
+        // For example:
+        // PaydayPoolReward to pool 1
+        // PaydayAccountReward to account 5
+        // PaydayAccountReward to account 6
+        // PaydayAccountReward to account 1
+        // PaydayPoolReward to pool 8
+        // PaydayAccountReward to account 8
+        // PaydayAccountReward to account 2
+        // PaydayPoolReward to `None`
+        // PaydayAccountReward to account 10
+        // PaydayAccountReward to account 3
+        // Means 5, 6 are receiving rewards from delegating to 1 and 2 receiving rewards
+        // from delegating to 8, and 10, 3 are receiving rewards from passive
+        // delegation.
+        //
+        // `PaydayPoolReward` and `PaydayAccountReward` events only occur in payday
+        // blocks starting in protocol 4.
+        let mut total_rewards_pool_owners: Vec<Option<i64>> = vec![];
+        let mut total_transaction_rewards: Vec<i64> = vec![];
+        let mut total_baking_rewards: Vec<i64> = vec![];
+        let mut total_finalization_rewards: Vec<i64> = vec![];
+
+        let mut delegators_rewards_pool_owners: Vec<Option<i64>> = vec![];
+        let mut delegators_rewards_canonical_addresses: Vec<Vec<u8>> = vec![];
+        let mut delegators_transaction_rewards: Vec<i64> = vec![];
+        let mut delegators_baking_rewards: Vec<i64> = vec![];
+        let mut delegators_finalization_rewards: Vec<i64> = vec![];
+
+        let mut last_pool_owner: Option<Option<i64>> = None;
+
+        for event in events {
+            match event {
+                SpecialTransactionOutcome::PaydayPoolReward {
+                    pool_owner,
+                    transaction_fees,
+                    baker_reward,
+                    finalization_reward,
+                } => {
+                    let pool_owner =
+                        pool_owner.map(|baker_id| baker_id.id.index.try_into()).transpose()?;
+
+                    total_rewards_pool_owners.push(pool_owner);
+                    total_transaction_rewards.push(transaction_fees.micro_ccd.try_into()?);
+                    total_baking_rewards.push(baker_reward.micro_ccd.try_into()?);
+                    total_finalization_rewards.push(finalization_reward.micro_ccd.try_into()?);
+
+                    last_pool_owner = Some(pool_owner);
+                }
+                SpecialTransactionOutcome::PaydayAccountReward {
+                    transaction_fees,
+                    baker_reward,
+                    finalization_reward,
+                    account,
+                } => {
+                    // Collect all rewards from the delegators and associate the rewards with their
+                    // baker pools.
+                    if let Some(last_pool_owner) = last_pool_owner {
+                        delegators_rewards_pool_owners.push(last_pool_owner);
+                        delegators_rewards_canonical_addresses
+                            .push(account.get_canonical_address().0.to_vec());
+                        delegators_transaction_rewards.push(transaction_fees.micro_ccd.try_into()?);
+                        delegators_baking_rewards.push(baker_reward.micro_ccd.try_into()?);
+                        delegators_finalization_rewards
+                            .push(finalization_reward.micro_ccd.try_into()?);
+                    }
+                }
+                _ => {}
+            }
+        }
+
+        Ok(Self {
+            block_height,
+            total_rewards_pool_owners,
+            total_transaction_rewards,
+            total_baking_rewards,
+            total_finalization_rewards,
+            delegators_rewards_pool_owners,
+            delegators_rewards_canonical_addresses,
+            delegators_transaction_rewards,
+            delegators_baking_rewards,
+            delegators_finalization_rewards,
+        })
+    }
+
+    async fn save(
+        &self,
+        tx: &mut sqlx::Transaction<'static, sqlx::Postgres>,
+    ) -> anyhow::Result<()> {
+        // Calculate and insert the delegators' rewards.
+        // Don't record the rewards if they are associated with the baker itself
+        // (not a delegator) hence we check that `pool_owner IS DISTINCT FROM
+        // account_index`.
+        sqlx::query!(
+            "
+            INSERT INTO bakers_payday_pool_rewards (
+                payday_block_height,
+                pool_owner,
+                payday_delegators_transaction_rewards,
+                payday_delegators_baking_rewards,
+                payday_delegators_finalization_rewards
+            )
+            SELECT
+                $1 AS payday_block_height,
+                pool_owner,
+                SUM(
+                    CASE WHEN 
+                        pool_owner IS DISTINCT FROM account_index
+                            THEN payday_delegators_transaction_rewards 
+                            ELSE 0 
+                    END
+                ) AS payday_delegators_transaction_rewards,
+                SUM(
+                    CASE WHEN 
+                        pool_owner IS DISTINCT FROM account_index 
+                            THEN payday_delegators_baking_rewards 
+                            ELSE 0 
+                    END
+                ) AS payday_delegators_baking_rewards,
+                SUM(
+                    CASE 
+                    WHEN pool_owner IS DISTINCT FROM account_index
+                        THEN payday_delegators_finalization_rewards 
+                        ELSE 0 
+                    END
+                ) AS payday_delegators_finalization_rewards
+            FROM (
+                SELECT 
+                    pool_owner_data.pool_owner, 
+                    accounts.index AS account_index,
+                    tx_rewards.payday_delegators_transaction_rewards,
+                    baker_rewards.payday_delegators_baking_rewards,
+                    final_rewards.payday_delegators_finalization_rewards
+                FROM 
+                    UNNEST($2::BIGINT[]) WITH ORDINALITY AS pool_owner_data(pool_owner, idx)
+                    JOIN UNNEST($3::BYTEA[]) WITH ORDINALITY AS addresses(canonical_address, \
+             idx_addr) ON idx = idx_addr
+                    LEFT JOIN accounts ON accounts.canonical_address = addresses.canonical_address
+                    JOIN UNNEST($4::BIGINT[]) WITH ORDINALITY AS \
+             tx_rewards(payday_delegators_transaction_rewards, idx_tx) ON idx = idx_tx
+                    JOIN UNNEST($5::BIGINT[]) WITH ORDINALITY AS \
+             baker_rewards(payday_delegators_baking_rewards, idx_baker) ON idx = idx_baker
+                    JOIN UNNEST($6::BIGINT[]) WITH ORDINALITY AS \
+             final_rewards(payday_delegators_finalization_rewards, idx_final) ON idx = idx_final
+            )
+            GROUP BY pool_owner;
+            ",
+            &self.block_height,
+            &self.delegators_rewards_pool_owners as &[Option<i64>],
+            &self.delegators_rewards_canonical_addresses,
+            &self.delegators_transaction_rewards,
+            &self.delegators_baking_rewards,
+            &self.delegators_finalization_rewards
+        )
+        .execute(tx.as_mut())
+        .await
+        .context("Failed inserting delegator rewards at payday block")?;
+
+        // Insert the total rewards.
+        sqlx::query!(
+            "
+            UPDATE bakers_payday_pool_rewards AS rewards
+            SET 
+                payday_total_transaction_rewards = data.payday_total_transaction_rewards,
+                payday_total_baking_rewards = data.payday_total_baking_rewards,
+                payday_total_finalization_rewards = data.payday_total_finalization_rewards
+            FROM (
+                SELECT
+                    UNNEST($1::BIGINT[]) AS pool_owner,
+                    UNNEST($2::BIGINT[]) AS payday_total_transaction_rewards,
+                    UNNEST($3::BIGINT[]) AS payday_total_baking_rewards,
+                    UNNEST($4::BIGINT[]) AS payday_total_finalization_rewards
+            ) AS data
+            WHERE rewards.pool_owner IS NOT DISTINCT FROM data.pool_owner
+            AND rewards.payday_block_height = $5
+            ",
+            &self.total_rewards_pool_owners as &[Option<i64>],
+            &self.total_transaction_rewards,
+            &self.total_baking_rewards,
+            &self.total_finalization_rewards,
+            &self.block_height,
+        )
+        .execute(tx.as_mut())
+        .await
+        .context("Failed inserting total rewards at payday block")?;
+
+        Ok(())
+    }
+}
+
 /// Insert special transaction outcomes for a particular block.
 struct PreparedInsertBlockSpecialTransacionOutcomes {
     /// Height of the block containing these special events.
-    block_height:        i64,
+    block_height: i64,
     /// Index of the outcome within this block in the order they
     /// occur in the block.
     block_outcome_index: Vec<i64>,
     /// The types of the special transaction outcomes in the order they
     /// occur in the block.
-    outcome_type:        Vec<SpecialEventTypeFilter>,
+    outcome_type: Vec<SpecialEventTypeFilter>,
     /// JSON serializations of `SpecialTransactionOutcome` in the order they
     /// occur in the block.
-    outcomes:            Vec<serde_json::Value>,
+    outcomes: Vec<serde_json::Value>,
+    /// The `SpecialEvents` of a payday block in the order they
+    /// occur in the block.
+    payday_special_transacion_outcomes: PreparedPaydaySpecialTransacionOutcomes,
 }
 
 impl PreparedInsertBlockSpecialTransacionOutcomes {
@@ -4282,6 +4508,10 @@ impl PreparedInsertBlockSpecialTransacionOutcomes {
         let mut block_outcome_index = Vec::with_capacity(events.len());
         let mut outcome_type = Vec::with_capacity(events.len());
         let mut outcomes = Vec::with_capacity(events.len());
+
+        let payday_special_transacion_outcomes =
+            PreparedPaydaySpecialTransacionOutcomes::prepare(block_height, events)?;
+
         for (block_index, event) in events.iter().enumerate() {
             let outcome_index = block_index.try_into()?;
             let special_event = SpecialEvent::from_special_transaction_outcome(
@@ -4298,6 +4528,7 @@ impl PreparedInsertBlockSpecialTransacionOutcomes {
             block_outcome_index,
             outcome_type,
             outcomes,
+            payday_special_transacion_outcomes,
         })
     }
 
@@ -4327,6 +4558,9 @@ impl PreparedInsertBlockSpecialTransacionOutcomes {
         .execute(tx.as_mut())
         .await?
         .ensure_affected_rows(self.outcomes.len().try_into()?)?;
+
+        self.payday_special_transacion_outcomes.save(tx).await?;
+
         Ok(())
     }
 }